requirements.txt
# Core dependencies for Navi Gym
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# Genesis physics engine dependencies
taichi>=1.7.0
trimesh>=3.9.0
Pillow>=8.0.0
pyyaml>=5.4.0

# RL and ML libraries
gymnasium>=0.28.0
stable-baselines3>=2.0.0

# Customer integration dependencies
websockets>=11.0.0
aiohttp>=3.8.0
fastapi>=0.68.0
uvicorn>=0.15.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.15.0

# Development and testing
pytest>=6.0.0
pytest-asyncio>=0.18.0
black>=21.0.0
flake8>=4.0.0
mypy>=0.910

# Asset processing
opencv-python>=4.5.0
imageio>=2.9.0
matplotlib>=3.5.0

# Audio processing (for TTS integration)
librosa>=0.8.0
soundfile>=0.10.0

# Utilities
tqdm>=4.62.0
rich>=10.0.0
click>=8.0.0
